refactor: use model in checkout repository and service
